+++ docs/changelog.rst
@@ -8,6 +8,11 @@ Changelog
 0.6.15 (unreleased)
 *******************
 
+Internal Changes
+================
+- Fix the license syntax in ``CITATION.cff`` to help Zenodo recognize the license type (:merge:`546`). By `Kyle
+  Brindley`_.
+
 *******************
 0.6.14 (2023-03-23)
 *******************