+++ docs/changelog.rst
@@ -12,6 +12,7 @@ Internal Changes
 ================
 - Fix the license syntax in ``CITATION.cff`` to help Zenodo recognize the license type (:merge:`546`). By `Kyle
   Brindley`_.
+- Added ``--exclude-regex`` argument to ``visualize`` subcommand (:issue:`419`, :merge:`548`). By `Sergio Cordova`_.
 - Cleaned up conda package CI files after ``conda build`` (:issue:`442`, :merge:`547`). By `Sergio Cordova`_.
 - Linked argparse tutorial in tutorial 01 (:issue:`439`, :merge:`549`). By `Sergio Cordova`_.