+++ docs/changelog.rst
@@ -19,6 +19,7 @@ Internal Changes
 - Fix the license syntax in ``CITATION.cff`` to help Zenodo recognize the license type (:merge:`546`). By `Kyle
   Brindley`_.
 - Cleaned up conda package CI files after ``conda build`` (:issue:`442`, :merge:`547`). By `Sergio Cordova`_.
+- Removed mutable default arguments from python scripts (:issue:`454`, :merge:`553`). By `Sergio Cordova`_.
 
 Enhancements
 ============