+++ docs/changelog.rst
@@ -28,6 +28,7 @@ Enhancements
   Matlab script(s) to build directory prior to execution. Matlab script copy operation no longer performed by default.
   Builder still considered "experimental" until a tutorial is released (:issue:`456`, :merge:`555`). By `Kyle
   Brindley`_.
+- Add an example Matlab input parser to the Matlab tutorial script (:issue:`420`, :merge:`556`). By `Kyle Brindley`_.
 
 *******************
 0.6.14 (2023-03-23)